feat: replace pmtiles with Maptiler tileset
Note that we add a Maptiler key to the repo here which is okay. It is highly restrictive to allow only selected origins.

Why the switch away from pmtiles: I was not able to make them work with GitHub Pages. There were errors like "Wrong magic number" from the pmtiles frontend lib or CORS issues when accessing the pmtiles from a separate GitHub Pages deployment. In the end, I want to move forward with this prototype and not debug obscure errors, so I have decided to just upload the tileset to Maptiler and call it a day. Docs to follow.